// apps/ui/src/client/model/sandboxSummaryState.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Nibiru API
 * Nibiru API Documentation
 * OpenAPI spec version: 1.0.0
 */

export type SandboxSummaryState = typeof SandboxSummaryState[keyof typeof SandboxSummaryState];


export const SandboxSummaryState = {
  creating: 'creating',
  running: 'running',
  stopping: 'stopping',
  stopped: 'stopped',
  starting: 'starting',
  deleting: 'deleting',
  failed: 'failed',
} as const;